atomic open symlink: Reply ok with the attribute, but no inode open
Atomic open needs to do a very special symlink handling - it
needs to reply with all the attributes, but at the same time
must not keep inode references - as if opening would have failed.
Kernel side is going to use the attributes, but does not expect
any further open actions on the userspace side.