FUSE_INIT handler: Set got_init only shortly before send_reply_ok
A previous commit added single thread behavior as long as se->got_init is
not set, especially as buffer allocation should be done after got_init
only. Technically it should not make a difference in current libfuse design,
as more worker threads are only allocated after the FUSE_INIT handler
is completely done. That might at some point change, though. In general
it is cleaner to set se->got_init, when the init handler is done.

Done as separate commit, as this introduces a slight risk to break
something - backport to older version should be done carefully.

Signed-off-by: Bernd Schubert <bernd@bsbernd.com>